♻️ [refactor] move mac address resolution to avoid duplication
The mac address resolution was done everytime we needed the mac
address, instead of doing it once at the beginning while loading
the configuration. This fixes that, and refactors the tests since
the corresponding utils function changed module.
